PYIC-7309: - Add accessibility check for template pages - Make browser tests run in parallel for performance
