+++ browser-tests/package.json
@@ -10,6 +10,7 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
+    "@axe-core/playwright": "^4.10.0",
     "@playwright/test": "1.46.0",
     "dotenv": "^16.4.5"
   }
